Testing that gif works
